pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi_template"
version = "0.1.0"
description = "A FastAPI template"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Clément Malige" }
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.28",
    "alembic>=1.13.1",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0", # For PostgreSQL
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9", # For form data handling
    "email-validator>=2.1.1",
    "python-dotenv>=1.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "black>=24.3.0",
    "coverage>=7.9.1",
    "isort>=5.13.2",
    "mypy>=1.9.0",
    "pytest>=8.0.2",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.2.1",
    "httpx>=0.27.0",
    "pre-commit>=3.6.2",
]

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff.per-file-ignores]
"alembic/env.py" = ["F403", "F401"]
"tests/conftest.py" = ["F403", "F401"]
